shared/requirements.txt
# Shared dependencies for Flov7 platform
supabase==2.3.0
pydantic==2.5.0
python-dotenv==1.0.0
redis==5.0.1
